xong CRUD lop hoc, show, active sidebar

+++ app/Http/Controllers/Admin/ClassController.php
@@ -9,6 +9,8 @@ use App\Models\Course;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Str;
+use App\Http\Requests\Admin\ClassRequest;
 
 class ClassController extends Controller
 {
@@ -19,8 +21,10 @@ class ClassController extends Controller
      */
     public function index()
     {
+
         $classes = ClassStudy::select([
             'id',
+            'slug',
             'name',
             'amount'
         ])
@@ -37,12 +41,13 @@ class ClassController extends Controller
      */
     public function create()
     {
+        $course = [];
+        $class = new ClassStudy();
         $courses = Course::select([
             'id',
             'title',
-        ])
-            ->get();
-        return view('admin.modules.classes.create', compact('courses'));
+        ])->get();
+        return view('admin.modules.classes.create', compact('courses', 'class', 'course'));
     }
 
     /**
@@ -51,17 +56,17 @@ class ClassController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(ClassRequest $request)
     {
-        $class_item = $request->except('_token');
 
+        $class_item = $request->except('_token');
         DB::beginTransaction();
         try {
             $class = ClassStudy::create([
-                'name'   => $class_item['name'],
-                'slug' => \Str::slug($class_item['name']),
+                'name'          => $class_item['name'],
+                'slug'          => Str::slug($class_item['name']),
                 'description'   => $class_item['description'],
-                'amount'   => $class_item['amount'],
+                'amount'        => $class_item['amount'],
             ]);
             if (isset($_POST['course_id'])) {
                 foreach ($_POST['course_id'] as $value) {
@@ -86,8 +91,12 @@ class ClassController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show($slug)
     {
+        $class = ClassStudy::where('slug', $slug)->first();
+        $course = $class->courses()->get();
+        $std = $class->users()->get();
+        return view('admin.modules.classes.show', compact('class', 'course', 'std'));
     }
 
     /**
@@ -98,14 +107,18 @@ class ClassController extends Controller
      */
     public function edit($id)
     {
-        $class = ClassStudy::find($id)->with('courses');
+        $courses = Course::select([
+            'id',
+            'title',
+        ])->get();
+        $class = ClassStudy::find($id);
         if ($class) {
-            // $courses = $class->courses()-get();
-            // dd($courses);
-            return view('admin.modules.classes.edit', compact('class', 'courses'));
+            $course = $class->courses()->get();
+            return view('admin.modules.classes.edit', compact('class','courses', 'course'));
         }
         return redirect(route('class.index'))
-            ->with('msg', 'Không tìm thấy lớp học này');
+            ->with('message', 'Không tìm thấy lớp học này')
+            ->with('type_alert', "danger");
     }
 
     /**
@@ -115,9 +128,36 @@ class ClassController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(ClassRequest $request, $id)
     {
         //
+        $message = 'Lớp học không tồn tại!';
+        $class = ClassStudy::find($id);
+        if ($class) {
+            $class->name        = $request->input('name');
+            $class->slug        = Str::slug($class->name);
+            $class->description = $request->input('description');
+            $class->amount      = $request->input('amount');
+            $class->save();
+            $message            = 'Cập nhật lớp học thành công';
+        }
+
+        try {
+            if (isset($_POST['course_id'])) {
+
+                $class_dettach = ClassStudy::find($class->id);
+                $class_dettach->courses()->detach();
+                foreach ($_POST['course_id'] as $value) {
+                    //Xử lý các phần tử được chọn
+                    $course = Course::find($value);
+                    $class->courses()->attach($course->id);
+                }
+            }
+        } catch (\Throwable $t) {
+            dd($t);
+        }
+        return redirect(route('class.index'))
+        ->with('message', $message);
     }
 
     /**
@@ -131,10 +171,22 @@ class ClassController extends Controller
         $class_id = $request->input('class_id', 0);
         if ($class_id)
         {
-            ClassStudy::destroy($class_id);
-            return redirect(route('class.index'))
-            ->with('message', "Xóa lớp học thành công!")
-            ->with('type_alert', "success");
+            $data = ClassStudy::find($class_id);
+            $name = $data->name;
+
+            if($data->users->count() > 0){
+                return redirect(route('class.index'))
+                ->with('message', "Không thể xóa! Đang có học sinh đăng kí lớp")
+                ->with('type_alert', "danger");
+            }
+            else{
+                $class_dettach = ClassStudy::find($class_id);
+                ClassStudy::destroy($class_id);
+                $class_dettach->courses()->detach();
+                return redirect(route('class.index'))
+                ->with('message', "Xóa thành công: ". $name )
+                ->with('type_alert', "success");
+            }
         }else {
             throw new ModelNotFoundException();
         }
